pyproject.toml
[project]
name = "dreamsculpt_be"
version = "0.0.3"
description = "Backend server for DreamSculpt"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.89",
    "fastapi>=0.116.1",
    "google-genai>=1.63.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "uvicorn>=0.35.0",
    "xai-sdk>=1.11.0",
]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend="uv_build"

[project.scripts]
start = "dreamsculpt_be.main:main"

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]

[project.optional-dependencies]
host_model = [
    "torch>=2.8.0",
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.0",
    "diffusers>=0.35.1",
    "gguf>=0.17.1",
    "huggingface-hub>=0.34.5",
    "numpy>=2.3.3",
    "protobuf>=6.33.2",
    "sentencepiece>=0.2.1",
    "transformers>=4.56.2",
]
